+++ src/authorizations/utils/permissions.ts
@@ -1,5 +1,6 @@
 import {Bucket, Permission} from 'src/types'
 import {CLOUD} from 'src/shared/constants'
+import {capitalize} from 'lodash'
 
 type PermissionTypes = Permission['resource']['type']
 
@@ -268,4 +269,39 @@ export const formatApiPermissions = (permissions, orgID) => {
     }
   })
   return apiPerms
+}
+
+export const generateDescription = apiPermissions => {
+  let generatedDescription = ''
+
+  if (apiPermissions.length > 2) {
+    const actions = []
+    apiPermissions.forEach(perm => {
+      actions.push(perm.action)
+    })
+    const isRead = actions.some(action => action === 'read')
+    const isWrite = actions.some(action => action === 'write')
+
+    if (isRead && isWrite) {
+      generatedDescription += `Read Multiple Write Multiple`
+    } else if (isRead) {
+      generatedDescription += `Read Multiple`
+    } else if (isWrite) {
+      generatedDescription += `Write Multiple`
+    }
+  } else {
+    apiPermissions.forEach(perm => {
+      if (perm.resource.name) {
+        generatedDescription += ` ${capitalize(perm.action)} ${
+          perm.resource.type
+        } ${perm.resource.name} `
+      } else {
+        generatedDescription += ` ${capitalize(perm.action)} ${
+          perm.resource.type
+        } `
+      }
+    })
+  }
+
+  return generatedDescription
 }
